feat: add supabase auth and sync skeleton
