+++ .env.example
@@ -1,0 +1,4 @@
+VITE_SUPABASE_URL=
+VITE_SUPABASE_ANON_KEY=
+# OAuth redirect URI to add in Supabase dashboard and providers
+# <SUPABASE_URL>/auth/v1/callback
