+++ package.json
@@ -17,7 +17,11 @@
     "framer-motion": "^12.23.12",
     "lucide-react": "^0.539.0",
     "react": "^19.1.1",
-    "react-dom": "^19.1.1"
+    "react-dom": "^19.1.1",
+    "@supabase/supabase-js": "^2.44.0",
+    "zustand": "^4.5.2",
+    "react-router-dom": "^6.28.0",
+    "localforage": "^1.10.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.32.0",
